update: PLANS, IPlan, TPlanType
